pyproject.toml
[project]
name = "Alcove"
version = "0.1.0"
description = "Alcove 物华弥新小助手py重写版"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "maafw>=4.5.6",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
